Fix circular dependency between models/training_workflow.py and routes/trainable_models.py
Extract file I/O helpers (_slug, _model_path, _read_model, _write_model,
get_trainable_models_dir) into vtsearch/models/trainable_model_store.py and
sync_labels_to_loaded_model into vtsearch/models/label_sync.py. This breaks
the cycle where training_workflow imported from trainable_models and vice versa.
Update all callers across routes and tests to use canonical import paths.
