+++ vtsearch/routes/medias.py
@@ -537,7 +537,7 @@ def vote_media(media_id: int) -> tuple[Response, int] | Response:
 
     toggle_vote(media_id, vote)
 
-    from vtsearch.routes.trainable_models import sync_labels_to_loaded_model
+    from vtsearch.models.label_sync import sync_labels_to_loaded_model
 
     sync_labels_to_loaded_model()
